+++ src/main/java/TMDTBoBa/BoBaEcor/Repository/User/IUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IUserRepository extends JpaRepository<User,Long> {
+public interface IUserRepository extends JpaRepository<User,Integer> {
     Optional<User> findByEmail(String email);
 
     Optional<User> findByPhoneNumber(String phone);
